final commit

+++ controllers/productController.js
@@ -1,32 +1,48 @@
 const Product = require("../models/productModel");
+const User = require("../models/userModel");
 
-//for add or fetch
+//for fetch
 exports.getProductController = async (req, res) => {
   try {
-    const products = await Product.find({});
+    const userId = req.user.id;
+    const products = await Product.find({ user: userId });
     res.status(200).json(products);
   } catch (error) {
     res.json("error fetching products");
   }
 };
 
-//for add
+//add
 exports.addProductController = async (req, res) => {
   try {
-    const newProducts = new Product(req.body);
-    await newProducts.save();
-    res.status(200).send("Products Created Successfully!");
+    const userId = req.user.id;
+    const newProduct = new Product({
+      ...req.body,
+      user: userId,
+    });
+    await newProduct.save();
+    const user = await User.findById(userId);
+    user.products.push(newProduct._id);
+    await user.save();
+
+    res.status(200).send("Product Created Successfully!");
   } catch (error) {
     console.log(error);
+    res.status(500).send("Server Error");
   }
 };
 
 //for update
 exports.updateProductController = async (req, res) => {
   try {
-    await Product.findOneAndUpdate({ _id: req.body.productId }, req.body, {
-      new: true,
-    });
+    const userId = req.user.id;
+    await Product.findOneAndUpdate(
+      { _id: req.body.productId, user: userId },
+      req.body,
+      {
+        new: true,
+      }
+    );
     res.status(201).json("Product Updated!");
   } catch (error) {
     res.status(400).send(error);
@@ -37,7 +53,8 @@ exports.updateProductController = async (req, res) => {
 //for delete
 exports.deleteProductController = async (req, res) => {
   try {
-    await Product.findOneAndDelete({ _id: req.body.productId });
+    const userId = req.user.id;
+    await Product.findOneAndDelete({ _id: req.body.productId, user: userId });
     res.status(200).json("Product Deleted!");
   } catch (error) {
     res.status(400).send(error);
